- moved Member_ForgotUsernameEmail.php to it's own file - moved code from code to src - started added references to classes
